+++ src/lightning/pytorch/core/mixins/hparams_mixin.py
@@ -15,6 +15,8 @@ import copy
 import inspect
 import types
 from argparse import Namespace
+from contextlib import contextmanager
+from contextvars import ContextVar
 from typing import Any, List, MutableMapping, Optional, Sequence, Union
 
 from lightning.fabric.utilities.data import AttributeDict
@@ -22,6 +24,18 @@ from lightning.pytorch.utilities.parsing import save_hyperparameters
 
 _PRIMITIVE_TYPES = (bool, int, float, str)
 _ALLOWED_CONFIG_TYPES = (AttributeDict, MutableMapping, Namespace)
+
+
+given_hyperparameters: ContextVar = ContextVar("given_hyperparameters", default=None)
+
+
+@contextmanager
+def given_hyperparameters_context(value):
+    token = given_hyperparameters.set(value)
+    try:
+        yield
+    finally:
+        given_hyperparameters.reset(token)
 
 
 class HyperparametersMixin:
@@ -105,12 +119,13 @@ class HyperparametersMixin:
 
         """
         self._log_hyperparams = logger
+        given_hparams = given_hyperparameters.get()
         # the frame needs to be created in this file.
-        if not frame:
+        if given_hparams is None and not frame:
             current_frame = inspect.currentframe()
             if current_frame:
                 frame = current_frame.f_back
-        save_hyperparameters(self, *args, ignore=ignore, frame=frame)
+        save_hyperparameters(self, *args, ignore=ignore, frame=frame, given_hparams=given_hparams)
 
     def _set_hparams(self, hp: Union[MutableMapping, Namespace, str]) -> None:
         hp = self._to_hparams_dict(hp)